requirements.txt
pandas==2.2.2
numpy==1.26.4
google-cloud-storage==2.16.0
google-cloud-bigquery==3.25.0
openpyxl==3.1.2